apps/backend/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
supabase==2.4.3
python-dotenv==1.0.1
